chore: move start kyc mutation hook to module
